Add `ag-compute` control.
This is an invisible control, it can watch for a value from another
control and apply some transformation to it.

The same text is used is different parts of a metric name but
in some cases it should be in upper case and in some cases it should
be in lower case. For example we query a metric: `some.namespace.FOO:metric.name-foo`

To build an expression we can define 2 parameters: `my_param` - a dropdown in uppercase
and `my_param_ls` is an invisible compute parameter. It takes value of the `my_param`
and converts it to a lower case:

```
<ag-select name="my_param" label="Very important parameter" default='FOO'>
  <option value="FOO">FOO</option>
  <option value="BAR">BAR</option>
  <option value="BAZ">BAZ</option>
</ag-select>
<ag-compute name="my_param_lc">{{my_param | lowercase}}</ag-compute>
```

and later both of the parameters can be applied to a metric expression:

```
$start$:$end$:some.namespace.$my_param$:metric.name-$my_param_lc$:avg
```