W-5821201: GUS/GOC INVALID_AUTH_HEADER error (#340)
Extracting GOCTransport (renamed to GusTransport) and EndpointInfo classes from the GOCNotifier class, so it can be used by the GusNotifier class.
Migrating GusNotifier class to use GusTransport for auth and getting token. Removing duplicate code.
Updating GusTransport class to use a LoadingCache to hold the current EndpointInfo and manage refreshes. Also refactored the logic to get token into a Supplier. Making it so that HttpClient is re-used instead of being one-time-use.
Re-using GusTransport instance in GusNotifier and GOCNotifier instead of being one-time use. Also adding in retries for GusNotifier and GOCNotifier if notification fails to post. If failure is due to bad token, the notifier will attempt to refresh the token cache.
Adding metrics to track number of retries and number of failures for GOCNotifier and GusNotifier.

@rev @nkarri @ikeck @nikhil-wahi @sudhansu-bahety
@W-5821201

https://gus.lightning.force.com/lightning/r/ADM_Work__c/a07B0000005jY7IIAU/view